feat(communities): isolate get controller
